+++ CHANGELOG.md
@@ -13,6 +13,8 @@
 - fix text contrast issues for accessibility
 - use virtual table
 - remove theme pre-load script
+- move computation to ingest worker
+- split API into summary and details
 
 ## [0.4.1] - 2025-10-24